feat(show-warning-manifest-file): added the show warning icon login at the manifest tab title
